Linux bash getopt wants equals for args, less *'s
On my Ubuntu 14.04 box, flashing was failing because getopt was taking
"--gaia PATH" and treating it as just "--gaia" with PATH moved to after the
"--" stuff.  It turns out, per my local "man getopt", that optional arguments
are requirement to have the form "--foo=bar", you can only use "--foo bar" if
it's a required argument.

Also, I noticed that the filenames had asterisks in them from nameBits
previously being used for regexp purposes.  Arguably it's preferable to leave
the wildcards out of the filenames.  It certainly makes them look a lot nicer.